pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tabularqual"
version = "0.1.0"
description = "Convert spreadsheets specified in SpreadSBML to SBML-qual"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "python-libsbml==5.20.4",
  "openpyxl==3.1.5",
  "packaging>=24.0",
  "click==8.1.7",
  "python-dateutil>=2.9.0",
]

[project.scripts]
to-sbml = "converter.cli:to_sbml_entry"

[tool.setuptools]
packages = ["converter"]